pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "dsb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "av>=16.1.0",
    "coverage>=7.13.1",
    "datasets<3.0.0",
    "diffusers>=0.36.0",
    "dotenv>=0.9.9",
    "dvc[gdrive]>=3.66.0",
    "einops>=0.8.1",
    "evaluate>=0.4.6",
    "ffmpeg>=1.4",
    "gdown>=5.2.0",
    "huggingface-hub>=1.3.3",
    "hydra-core==1.3.2",
    "invoke>=2.2.1",
    "ipykernel>=7.1.0",
    "librosa>=0.11.0",
    "matplotlib==3.10.8",
    "mnist==0.2.2",
    "nd2>=0.11.2",
    "nemo-toolkit[asr]>=2.1.0",
    "numpy==2.4.0",
    "omegaconf==2.3.0",
    "onnxruntime-gpu>=1.24.2",
    "opencv-python>=4.13.0.92",
    "pre-commit>=4.5.1",
    "pytorch-lightning==2.6.0",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "rotary-embedding-torch>=0.8.9",
    "scipy>=1.17.1",
    "soundfile>=0.13.1",
    "timm>=1.0.24",
    "torch==2.9.1",
    "torch-ema==0.3",
    "torchaudio>=2.9.1",
    "torchview>=0.2.7",
    "torchvision>=0.24.1",
    "transformers>=5.2.0",
    "types-requests>=2.32.4.20250913",
    "wandb[media]==0.23.0",
    "xlrd>=2.0.2",
    "zarr>=3.1.5",
]

[dependency-groups]
dev = [
    "coverage>=7.13.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]

[tool.ruff.lint.isort]
known-third-party = ["wandb", "hydra"]

[tool.mypy]
disable_error_code = ["override", "assignment"]
ignore_missing_imports = true

[tool.ruff]
line-length=120
